refactor isReactCompositeComponent
