--- src/utils/isReactCompositeComponent.ts
@@ -1,12 +1,0 @@
-export default function isReactCompositeComponent(type: unknown): boolean {
-  if (typeof type !== 'function') {
-    return false;
-  }
-  if (typeof type.prototype !== 'object') {
-    return false;
-  }
-  if (typeof type.prototype.render !== 'function') {
-    return false;
-  }
-  return true;
-}
